Exercise 11
Remove the C# code that updates the Celsius text box when the slider changes.
In MainWindow.xaml, create a CelsiusConverter resource and name it celsiusConverter. In order to reference your own classes, you need to specify to which namespace it they belong. Notice xmlns:local="clr-namespace:View" on line 6: this introduces a shorthand way to refer to classes in the View namespace (to which your CelsiusConverter belongs). In other words, use local:CelsiusConverter to refer from within the XAML file to your CelsiusConverter class.
Bind the Celsius text box's Text property to the slider's Value property.
Have the binding update itself as soon as the user modifies the text box's contents.
Tell the binding to use celsiusConverter as a converter. Note that we need to refer to the object, not the class.