Exercise 17
Refactoring - Create a new file ViewModel.cs.
In this file, create a new class ConverterViewModel.
Move all TemperatureInKelvin related code to this class. This includes the field, the actual property, the INotifyPropertyChanged interface.
Remove the INotifyPropertyChanged functionality from the MainWindow class.
In the MainWindow constructor, set the MainWindows DataContext to a new instance of ConverterViewModel.