Exercise 20
MVVM - To achieve this separation:

In the solution explorer, right click on the solution TemperatureConverter (not the project).
In this context menu, choose "New Project".
From the list, pick Class Library (.NET Framework). Make sure you pick the one mentioning ".NET Framework".
Name it Model.
Visual Studio automatically adds a file, probably named Class1.cs. Remove it.
Repeat these steps, but use ViewModel as name.
TemperatureScale.cs currently resides in the View-project. Copy (drag and drop) it to the Model project and remove it from the View project.
Analogously, move ViewModel.cs to the ViewModel project.
Right-click on ViewModel and choose to add references. Add a check mark to the Model project, to tell Visual Studio that ViewModel depends on Model.
View depends on ViewModel. Add a reference.