Exercise 28
We only want to allow temperatures ranging from 0K to 1000K. Have CanExecute change to false whenever incrementing/decrementing would cause the temperature to move outside this range.

Don't forget to deal with CanExecuteChanged! You can check if it works by moving the slider: position it to the left and the - button should become grayed out. Then move it to the right: - should be enabled and + disabled.